imp: further improvements
